+++ src/server/db/queries.ts
@@ -1,0 +1,36 @@
+import { eq } from "drizzle-orm";
+
+import { db } from "~/server/db";
+import {
+  files_table as fileSchema,
+  folders_table as folderSchema,
+} from "~/server/db/schema";
+
+export async function getAllParentsForFolder(folderId: number) {
+  const parents = [];
+  let currentId: number | null = folderId;
+  while (currentId != null) {
+    const folders = await db
+      .selectDistinct()
+      .from(folderSchema)
+      .where(eq(folderSchema.id, currentId));
+
+    if (!folders[0]) throw new Error("parent folder not found");
+
+    parents.unshift(folders[0]);
+    currentId = folders[0]?.parent; // parent can be null
+  }
+
+  return parents;
+}
+
+export function getAllFolders(folderId: number) {
+  return db
+    .select()
+    .from(folderSchema)
+    .where(eq(folderSchema.parent, folderId));
+}
+
+export function getAllFiles(folderId: number) {
+  return db.select().from(fileSchema).where(eq(fileSchema.parent, folderId));
+}
